[Complex] : UI 위치 조정 및 골드 다이아몬드 재화 자료형 long으로 변경

+++ Assets/02.Scripts/02.Data/00.Loader/UserDB.cs
@@ -116,12 +116,12 @@ public class UserDB
     /// <summary>
     /// Gold
     /// </summary>
-    public int Gold;
+    public long Gold;
 
     /// <summary>
     /// Diamonds
     /// </summary>
-    public int Diamonds;
+    public long Diamonds;
 
     /// <summary>
     /// PlayTimeInSeconds
